move publish from generation
